Make authenticate_secondarily a util
